backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
scikit-learn==1.3.2
numpy==1.24.3
shap==0.43.0
reportlab==4.0.7
cryptography==41.0.7
python-dotenv==1.0.0
pyjwt==2.8.1
passlib==1.7.4
bcrypt==4.1.1
